Downloads/mlfinal/anbu/backend/requirements.txt
numpy
pandas
scikit-learn
django
djangorestframework
django-cors-headers